requirements.txt
amqp
asgiref
attrs
autobahn
Automat
billiard
cbor2
celery
certifi
cffi
channels
channels_redis
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
constantly
cron_descriptor
cryptography
daphne
Django
django-celery-beat
django-timezone-field
djangorestframework
djangorestframework_simplejwt
google-auth
hyperlink
idna
Incremental
kombu
msgpack
packaging
pillow
prompt_toolkit
psycopg2
psycopg2-binary
py-ubjson
pyasn1
pyasn1_modules
pycparser
PyJWT
pyOpenSSL
python-crontab
python-dateutil
python-decouple
redis
requests
rsa
service-identity
six
sqlparse
Twisted
txaio
typing_extensions
tzdata
tzlocal
ujson
urllib3
vine
wcwidth
zope.interface

